Coverage 1/?

+++ src/objectives/extrema.jl
@@ -4,5 +4,5 @@ Computes the distance between extrema in an ordered set.
 """
 function dist_extrema(values::T...) where {T <: Number}
     m, M = extrema(values)
-    return M - m
+    return Float64(M - m)
 end
